Move version from rsd package to setup.py
I do not like calling exec() in setup.py and I do not see why the rsd
package needs a __version__ separate from the version of the project
distribution, considering that any dependency on a version of
reciprocal_smallest_distance should be on the distribution.  Examples of
dependencies would be in a pip requirements.txt file, or in a setup.py
install_requires argument, both of which depend on pypi packages (aka
distributions), not python modules/packages.