Move fasta.py to tfd. Relax fasta parsing.
The version was bumped and distribution uploaded to pypi.  This should
fix the __version__ bug (Issue 1).

Also fasta parsing was relaxed to ignore all blank lines.  This should
fix github Issue 1 as well.

Module fasta.py was moved to the tfd package and added as a dependency
in setup.py.